SG-004-v2: Replace 833-line swarm copies with 100-line interfaces (#18)
SG-004 copied 4 swarm entity files (agent.ts, task.ts, and their
repository interfaces) into CLI. These 833 lines would silently
drift when upstream changes swarm types.

Replace with swarm-interfaces.ts: 100 lines of minimal interfaces
matching only the properties CLI actually uses (id, name, status,
role, domain, getUtilization, etc). No drift risk — if swarm adds
new fields, they only need adding here when CLI needs them.
